requirements.txt
anthropic==0.8.1
python-dotenv==1.0.0
PyPDF2==3.0.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
requests==2.31.0 